client: Move response to each core API module
