+++ iota-client/core/add_neighbors.rs
@@ -22,3 +22,27 @@ pub(crate) fn add_neighbors(
         .body(body.to_string())
         .send()
 }
+
+/// This is a typed representation of the JSON response
+#[derive(Clone, Default, Debug, Serialize, Deserialize)]
+pub struct AddNeighborsResponse {
+    #[serde(rename = "addedNeighbors")]
+    added_neighbors: Option<usize>,
+    error: Option<String>,
+    exception: Option<String>,
+}
+
+impl AddNeighborsResponse {
+    /// Returns the added neighbors
+    pub fn added_neighbors(&self) -> &Option<usize> {
+        &self.added_neighbors
+    }
+    /// Returns the error attribute
+    pub fn error(&self) -> &Option<String> {
+        &self.error
+    }
+    /// Returns the exception attribute
+    pub fn exception(&self) -> &Option<String> {
+        &self.exception
+    }
+}
