+++ iota-client/core/get_neighbors.rs
@@ -1,6 +1,7 @@
 use reqwest::r#async::{Client, Response};
 use reqwest::Error;
 use tokio::prelude::Future;
+use iota_model::Neighbor;
 
 /// Returns the set of neighbors you are connected with, as
 /// well as their activity count. The activity counter is reset
@@ -19,4 +20,24 @@ pub(crate) fn get_neighbors(
         .header("X-IOTA-API-Version", "1")
         .body(body.to_string())
         .send()
+}
+
+/// This is a typed representation of the JSON response
+#[derive(Clone, Default, Serialize, Deserialize, Debug)]
+pub struct GetNeighborsResponse {
+    /// Any errors that occurred
+    error: Option<String>,
+    /// Neighbors if found
+    neighbors: Option<Vec<Neighbor>>,
+}
+
+impl GetNeighborsResponse {
+    /// Returns the error attribute
+    pub fn error(&self) -> &Option<String> {
+        &self.error
+    }
+    /// Returns the neighbors attribute
+    pub fn neighbors(self) -> Option<Vec<Neighbor>> {
+        self.neighbors
+    }
 }
