+++ iota-client/core/mod.rs
@@ -16,3 +16,18 @@ pub(crate) mod interrupt_attaching_to_tangle;
 pub(crate) mod remove_neighbors;
 pub(crate) mod store_transactions;
 pub(crate) mod were_addresses_spent_from;
+
+pub(crate) use add_neighbors::AddNeighborsResponse;
+pub(crate) use attach_to_tangle::AttachToTangleResponse;
+pub(crate) use broadcast_transactions::BroadcastTransactionsResponse;
+pub(crate) use find_transactions::FindTransactionsResponse;
+pub(crate) use get_balances::GetBalancesResponse;
+pub(crate) use get_inclusion_states::GetInclusionStatesResponse;
+pub(crate) use get_neighbors::GetNeighborsResponse;
+pub(crate) use get_node_info::GetNodeInfoResponse;
+pub(crate) use get_tips::GetTipsResponse;
+pub(crate) use get_transactions_to_approve::GetTransactionsToApprove;
+pub(crate) use get_trytes::GetTrytesResponse;
+pub(crate) use remove_neighbors::RemoveNeighborsResponse;
+pub(crate) use store_transactions::StoreTransactionsResponse;
+pub(crate) use were_addresses_spent_from::WereAddressesSpentFromResponse;
