+++ iota-client/core/remove_neighbors.rs
@@ -23,3 +23,34 @@ pub(crate) fn remove_neighbors(
         .body(body.to_string())
         .send()
 }
+
+/// This is a typed representation of the JSON response
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct RemoveNeighborsResponse {
+    /// Any errors that occurred
+    error: Option<String>,
+    /// Any exceptions that occurred
+    exception: Option<String>,
+    /// Amount of neighbors removed
+    #[serde(rename = "removedNeighbors")]
+    removed_neighbors: Option<usize>,
+}
+
+impl RemoveNeighborsResponse {
+    /// Returns the error attribute
+    fn error(&self) -> &Option<String> {
+        &self.error
+    }
+    /// Returns the exception attribute
+    fn exception(&self) -> &Option<String> {
+        &self.exception
+    }
+    /// Returns a reference to the amount of removed neighbors
+    fn removed_neighbors(&self) -> &Option<usize> {
+        &self.removed_neighbors
+    }
+    /// Consumes the response and returns the amount of removed neighbors if any
+    fn take_removed_neighbors(self) -> Option<usize> {
+        self.removed_neighbors
+    }
+}
