+++ iota-client/core/were_addresses_spent_from.rs
@@ -20,3 +20,27 @@ pub(crate) fn were_addresses_spent_from(
         .body(body.to_string())
         .send()
 }
+
+/// This is a typed representation of the JSON response
+#[derive(Clone, Serialize, Default, Deserialize, Debug)]
+pub struct WereAddressesSpentFromResponse {
+    /// Any errors that occurred
+    error: Option<String>,
+    /// States of addresses if found
+    states: Option<Vec<bool>>,
+}
+
+impl WereAddressesSpentFromResponse {
+    /// Returns the error attribute
+    fn error(&self) -> &Option<String> {
+        &self.error
+    }
+    /// Returns the states attribute
+    pub fn states(self) -> Option<Vec<bool>> {
+        self.states
+    }
+    /// Returns a specfic index into the states attribute
+    pub fn state(self, index: usize) -> bool {
+        self.states.unwrap_or_default()[index]
+    }
+}
